separate vue logic in specific file
